Changes bytes into Integers in Guess class.

+++ src/uk/ac/gcu/battleships/Guess.java
@@ -1,20 +1,20 @@
 package uk.ac.gcu.battleships;
 
 public class Guess {
-	private byte x_coord, y_coord;
+	private int x_coord, y_coord;
 	
 	//CONSTRUCTORS
 	
-	public Guess(byte x, byte y){
+	public Guess(int x, int y){
 		x_coord=x;
 		y_coord=y;
 	}
 	//METHODS
-	public byte get_X(){
+	public int get_X(){
 		return x_coord;
 	}
 	
-	public byte get_Y(){
+	public int get_Y(){
 		return y_coord;
 	}
 }
